% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_boot.R
\name{ps_boot}
\alias{ps_boot}
\title{Obtain a single bootstrap resample.}
\usage{
ps_boot(ps_object)
}
\arguments{
\item{ps_object}{An existing phyloseq object containing at least sample_data and otu_table elements where the taxa are assumed to be rows.}
}
\value{
A bootstrap resampled phyloseq object
}
\description{
Obtain a single bootstrap resample.
}
