% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_phyloseq.R
\name{resample_phyloseq}
\alias{resample_phyloseq}
\title{Perform bootstrap resampling of an existing phyloseq object.}
\usage{
resample_phyloseq(
  ps_object = ps,
  resamples = 1000,
  cores = 1,
  future_seed = 123
)
}
\arguments{
\item{ps_object}{An existing phyloseq object}

\item{resamples}{The number of bootstrap samples to generate (default = 1000)}

\item{cores}{Number of available cores}

\item{future_seed}{Seed for reproducing results (passed to future.seed)}
}
\value{
A list of bootstrap resampled phyloseq objects
}
\description{
The function assumes the phyloseq object has at least sample_data and otu_table elements and that taxa are rows. If your taxa are columns, transpose before running function.
}
